Switch to Symfony Flex for managing bundles
Behat is removed (for now) as this can be installed later using Flex, with the correct configuration and structure!

The AppKernel (in src/) uses the App namespace for compatibility with Symfony, and this
doesn’t really matter here as this is only used for development.

Some components are pre-installed as they will be used shortly.