Remove Doctrine ORM, use DBAL only
Using a full fledged CQRS/ES solution removes the need to use an ORM system,
secondly Doctrine ORM was already planed to be replaced with Pomm (for Postgres).

Note that Pomm will still planned for ReadModels, but not for creating projections (as this will require a custom EventStore,
to handle transactions properly).