[Webhosting] Fix Capabilities registering/guarding
It’s not possible to use an array as service-tag attribute value, instead
this is reduced configuring what really matters.
Templates now follow a convention.

The CommandToCapabilitiesGuard was wrongfully implemented, commands
trigger Capability limitations, and therefor a Command covers a capability instead
of a capability affecting a command. Plus, this makes the implementation easier.