[Security] Speed-up SplitToken usage
Don’t hide the Libsodium details in the SplitToken class but use a factory for generating,
this allows to use a Fake generator which speeds-up dependent tests.

The SplitTokenValueHolder no longer “depends” on the SplitToken class (to remove a Cyrillic Dependency)
and all validating is now performed by SplitToken directly.