[UI] Fix MessageFormType not handling violation
The ValidationMiddleware is handled before the Message handler is located
and executed, so the ValidationFailedException is part of the main exception
flow and not wrapped in ExceptionStack by the HandleMessageMiddleware.

This should actually be solved by using the ValidationMiddleware directly
rather than expecting this is how it works. But for now that's too much work.