[Webhosting] Add Suspension level (handling)
A Space can be suspended for a period of time, while retaining
the registration and all of its data.

When a Space is marked as Suspended it's still accessible but in read-only
modes. Depending on the Suspension level, FTP may still be accessible.

Commands or view-access might be blocked depending the suspension level.