[UI] Allow impersonating a user
This feature is available for admins only, and restricts admins from
personating other admins.

This commit also contains a fix for the `access_decision_manager.strategy`
that could abusively still grant access when another voter denied access, which
happens when you are a super-admin and want to impersonate (switch to) another
admin, the voter denies access but when the super-admin voter grants access
this was overruled. Instead the super-admin voter is executed as soon as is possible
which is after the `AuthenticatedVoter` and SwitchUserVoter. But before other voters.